requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
dnspython==2.8.0
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
Werkzeug==3.1.8
